requirements.txt
ccxt>=4.1.0
pandas>=2.2.0
numpy>=1.26.0
ta>=0.10.2
python-binance>=1.0.19
websocket-client>=1.6.4
python-dotenv>=1.0.0
schedule>=1.2.0
requests>=2.31.0
openpyxl>=3.1.0
pytz>=2023.3
